antlr4-for-python/hello-world/pyproject.toml
[project]
name = "hello-world"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "antlr4-python3-runtime>=4.13.2",
]
